start to move authentication logic to UserSession.
