+++ app/models/user.rb
@@ -53,17 +53,6 @@ class User < ActiveRecord::Base
     GoogleDrive.new(self)
   end
 
-  def self.authenticate(username, password)
-    user = User.find_by(
-      "email = :email OR username = :username",
-      username: username.downcase,
-      email: username.downcase
-    )
-    if user.present?
-      user.authenticate(password)
-    end
-  end
-
   private
 
   def create_profile
